tools/ocaml: fix 32-bit truncation bug in stub_xc_domain_setmaxmem()
Observed when attempting to boot a >4TiB VM:

After a call to domain_setmaxmem with 6370254848 (KiB), the domain's maxmem
got set to 2075287552, which is exactly 2^32 smaller.

xc_domain_setmaxmem takes an uint64_t as a parameter, and the OCaml value
is 64-bit already, so fix the C variable to match the type and avoid the
truncation.

Fixes: f5b43e95facd ("libxl: fix "xl mem-set" regression from 0c029c4da2")
Acked-by: Christian Lindig <christian.lindig@cloud.com>

+++ tools/ocaml/libs/xc/xenctrl_stubs.c
@@ -862,7 +862,7 @@ CAMLprim value stub_xc_domain_setmaxmem(value xch, value domid,
 	int retval;
 
 	uint32_t c_domid = _D(domid);
-	unsigned int c_max_memkb = Int64_val(max_memkb);
+	uint64_t c_max_memkb = Int64_val(max_memkb);
 	caml_enter_blocking_section();
 	retval = xc_domain_setmaxmem(_H(xch), c_domid,
 	                                 c_max_memkb);
